test(pnpr): move journal roll-forward tests to an external module
The pnpr dylint rule set rejects inline `#[cfg(test)] mod tests { ... }`
blocks, so the drop_conflicted_versions tests move to journal/tests.rs,
matching the crate's other modules.
